requirements.txt
fastapi==0.135.2
uvicorn==0.42.0
pydantic==2.12.5
fasttext-wheel==0.9.2
pandas==2.3.3
scikit-learn==1.8.0
requests==2.32.5
pyTelegramBotAPI==4.33.0
